rpcclient: handle backend versioning in one file
